+++ odoo/local-src/specific_report/models/round_instance.py
@@ -3,6 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 import math
+from collections import OrderedDict
 
 from odoo import api, models
 
@@ -32,3 +33,90 @@ class RoundInstance(models.Model):
             hour += 1
 
         return pattern % (hour, min)
+
+    @api.multi
+    def get_merged_shippings(self):
+        self.ensure_one()
+
+        shippings = self.shipping_ids.filtered(
+            lambda shipping: shipping.state == 'done'
+        )
+
+        shipping_values = {}
+        for shipping in shippings:
+            partner_value = shipping_values.get(shipping.partner_id, {})
+
+            number_of_drug = partner_value.get('number_of_drug', 0)
+            number_of_drug += shipping.number_of_drug
+            item_number_of_drug = partner_value.get('item_number_of_drug', 0)
+            item_number_of_drug += shipping.item_number_of_drug
+
+            number_of_cold = partner_value.get('number_of_cold', 0)
+            number_of_cold += shipping.number_of_cold
+            item_number_of_cold = partner_value.get('item_number_of_cold', 0)
+            item_number_of_cold += shipping.item_number_of_cold
+
+            number_of_food = partner_value.get('number_of_food', 0)
+            number_of_food += shipping.number_of_food
+            item_number_of_food = partner_value.get('item_number_of_food', 0)
+            item_number_of_food += shipping.item_number_of_food
+
+            number_of_equipment = partner_value.get('number_of_equipment', 0)
+            number_of_equipment += shipping.number_of_equipment
+            item_number_of_equipment = \
+                partner_value.get('item_number_of_equipment', 0)
+            item_number_of_equipment += shipping.item_number_of_equipment
+
+            number_of_human_drug = partner_value.get('number_of_human_drug', 0)
+            number_of_human_drug += shipping.number_of_human_drug
+            item_number_of_human_drug = \
+                partner_value.get('item_number_of_human_drug', 0)
+            item_number_of_human_drug += shipping.item_number_of_human_drug
+
+            number_total = partner_value.get('number_total', 0)
+            number_total += shipping.number_total
+            item_number_total = partner_value.get('item_number_total', 0)
+            item_number_total += shipping.item_number_total
+
+            note = partner_value.get('note', '')
+            if shipping.note:
+                note += " - %s" % shipping.note
+
+            partner_value.update({
+                'number_of_drug': number_of_drug,
+                'item_number_of_drug': item_number_of_drug,
+                'number_of_cold': number_of_cold,
+                'item_number_of_cold': item_number_of_cold,
+                'number_of_food': number_of_food,
+                'item_number_of_food': item_number_of_food,
+                'number_of_equipment': number_of_equipment,
+                'item_number_of_equipment': item_number_of_equipment,
+                'number_of_human_drug': number_of_human_drug,
+                'item_number_of_human_drug': item_number_of_human_drug,
+                'number_total': number_total,
+                'item_number_total': item_number_total,
+                'note': note,
+                'rank': shipping.rank,
+                'shipping': shipping,
+            })
+            shipping_values[shipping.partner_id] = partner_value
+
+        ordered_values = OrderedDict(
+            sorted(shipping_values.items(), key=lambda t: t[1].get('rank'))
+        )
+        result = []
+        for partner, values in ordered_values.iteritems():
+            shipping_value = shipping_values.get(partner)
+            if not shipping_value:
+                continue
+
+            # There is something very stupid in Odoo. If you want to display
+            # the address of a partner with the tag <address t-field=.... />
+            # you HAVE TO have at least one dot in the t-field
+            # (eg: t-field="shipping.partner_id" and not t-field="partner")
+            # It's why I append a shipping
+            result.append(
+                (partner, shipping_value['shipping'], shipping_value)
+            )
+
+        return result