+++ odoo/local-src/specific_stock/models/stock_pack_operation.py
@@ -8,3 +8,6 @@ class StockPackOperation(models.Model):
     location_name = fields.Char('Location name',
                                 related='location_id.name',
                                 store=True)
+    operator_id = fields.Many2one('res.users',
+                                  related='picking_id.operator_id',
+                                  readonly=True)